Refactor: put hpdyad logic in its own file
