+++ tests/benchmark_gamma.c
@@ -1,3 +1,9 @@
+/*
+ * SPDX-FileCopyrightText: 2025 Jonathan Busse <jonathan@jbusse.de>
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ */
+
 // SPDX-FileCopyrightText: 2024 Jonathan Busse <jonathan@jbusse.de>
 //
 // SPDX-License-Identifier: AGPL-3.0-only
