+++ tests/test_epsteinZeta.wls.license
@@ -1,0 +1,3 @@
+SPDX-FileCopyrightText: 2025 Jonathan Busse <jonathan@jbusse.de>
+
+SPDX-License-Identifier: AGPL-3.0-only
